requirements.txt
customtkinter
requests
beautifulsoup4
pandas
lxml
python-docx
lightgbm
scikit-learn
tqdm
